\name{spagmix-internal}
\Rdversion{1.1}
\alias{spagmix-internal}
\alias{rw}
\alias{riw}
\title{
Internal \code{spagmix} functions
}
\description{
Internal \code{\link{spagmix}} functions, only meant to be accessed by the user directly in a life-threatening situation.
}
\section{Usage}{
\code{rw(a,A)}\cr
\code{riw(b,B)}\cr
}
\keyword{internal}

\author{T.M. Davies}
